+++ lib/ui/color.dart
@@ -8,4 +8,8 @@ extension ColorX<T extends Widget> on T {
   Container backgroundColorInt(int color) {
     return Container(child: this, color: Color(color));
   }
+
+  linearGradientColor(List<Color> colors) {
+    return Container(child: this, decoration: BoxDecoration(gradient: LinearGradient(colors: colors)));
+  }
 }